cmd/pusher/main_test.go
package main

import (
    "bytes"
    "encoding/json"
    "net/http"
    "net/http/httptest"
    "testing"

    mqtt "github.com/eclipse/paho.mqtt.golang"
    "github.com/stretchr/testify/assert"
    "github.com/stretchr/testify/mock"
)

// MockMQTTClient is a mock implementation of the MQTT client
type MockMQTTClient struct {
    mock.Mock
}

func (m *MockMQTTClient) IsConnected() bool {
    args := m.Called()
    return args.Bool(0)
}

func (m *MockMQTTClient) Publish(topic string, qos byte, retained bool, payload interface{}) mqtt.Token {
    args := m.Called(topic, qos, retained, payload)
    return args.Get(0).(mqtt.Token)
}

// MockToken is a mock implementation of the MQTT token
type MockToken struct {
    mock.Mock
}

func (m *MockToken) Wait() bool {
    args := m.Called()
    return args.Bool(0)
}

func (m *MockToken) WaitTimeout(timeout time.Duration) bool {
    args := m.Called(timeout)
    return args.Bool(0)
}

func (m *MockToken) Error() error {
    args := m.Called()
    return args.Error(0)
}

func TestHandleSendMessage(t *testing.T) {
    // Initialize mock MQTT client
    mockMQTTClient := new(MockMQTTClient)
    mqttClient = mockMQTTClient

    // Test cases
    testCases := []struct {
        name           string
        payload        map[string]string
        expectedStatus int
        mqttError      error
    }{
        {
            name:           "Valid message",
            payload:        map[string]string{"content": "Test message"},
            expectedStatus: http.StatusOK,
            mqttError:      nil,
        },
        {
            name:           "Empty message",
            payload:        map[string]string{"content": ""},
            expectedStatus: http.StatusBadRequest,
            mqttError:      nil,
        },
        {
            name:           "MQTT publish error",
            payload:        map[string]string{"content": "Test message"},
            expectedStatus: http.StatusInternalServerError,
            mqttError:      assert.AnError,
        },
    }

    for _, tc := range testCases {
        t.Run(tc.name, func(t *testing.T) {
            // Create a mock token
            mockToken := new(MockToken)
            mockToken.On("Wait").Return(true)
            mockToken.On("Error").Return(tc.mqttError)

            // Set up expectations for the mock MQTT client
            mockMQTTClient.On("Publish", mock.Anything, mock.Anything, mock.Anything, mock.Anything).Return(mockToken)

            // Create a request body
            body, _ := json.Marshal(tc.payload)
            req, err := http.NewRequest("POST", "/send", bytes.NewBuffer(body))
            assert.NoError(t, err)

            // Create a ResponseRecorder to record the response
            rr := httptest.NewRecorder()

            // Call the handler
            handler := http.HandlerFunc(handleSendMessage)
            handler.ServeHTTP(rr, req)

            // Check the status code
            assert.Equal(t, tc.expectedStatus, rr.Code)

            // Verify mock expectations
            mockMQTTClient.AssertExpectations(t)
            mockToken.AssertExpectations(t)
        })
    }
}

func TestHandleHealthCheck(t *testing.T) {
    // Initialize mock MQTT client
    mockMQTTClient := new(MockMQTTClient)
    mqttClient = mockMQTTClient

    // Test cases
    testCases := []struct {
        name           string
        isConnected    bool
        expectedStatus int
    }{
        {
            name:           "MQTT connected",
            isConnected:    true,
            expectedStatus: http.StatusOK,
        },
        {
            name:           "MQTT disconnected",
            isConnected:    false,
            expectedStatus: http.StatusServiceUnavailable,
        },
    }

    for _, tc := range testCases {
        t.Run(tc.name, func(t *testing.T) {
            // Set up expectations for the mock MQTT client
            mockMQTTClient.On("IsConnected").Return(tc.isConnected)

            // Create a request
            req, err := http.NewRequest("GET", "/health", nil)
            assert.NoError(t, err)

            // Create a ResponseRecorder to record the response
            rr := httptest.NewRecorder()

            // Call the handler
            handler := http.HandlerFunc(handleHealthCheck)
            handler.ServeHTTP(rr, req)

            // Check the status code
            assert.Equal(t, tc.expectedStatus, rr.Code)

            // Verify mock expectations
            mockMQTTClient.AssertExpectations(t)
        })
    }
}

func TestPrometheusMiddleware(t *testing.T) {
    // Create a test handler
    testHandler := http.HandlerFunc(func(w http.ResponseWriter, r *http.Request) {
        w.WriteHeader(http.StatusOK)
    })

    // Wrap the test handler with the Prometheus middleware
    handler := prometheusMiddleware(testHandler)

    // Create a request
    req, err := http.NewRequest("GET", "/test", nil)
    assert.NoError(t, err)

    // Create a ResponseRecorder to record the response
    rr := httptest.NewRecorder()

    // Call the handler
    handler.ServeHTTP(rr, req)

    // Check the status code
    assert.Equal(t, http.StatusOK, rr.Code)

    // TODO: Add assertions to check if Prometheus metrics were updated
    // This would require exposing the metrics or using a test registry
}
